Change addingReationships Remove goals and favourite quote

+++ CharactersManager/Models/AppearanceViewModel.cs
@@ -3,7 +3,7 @@
     public class AppearanceViewModel
     {
         public int BreedId { get; set; }
-        public double Tall { get; set; }
+        public int Tall { get; set; }
         public string BodyType { get; set; }
         public string Description { get; set; }
     }
