+++ Service/Models/Appearance.cs
@@ -6,7 +6,7 @@ namespace Service.Models
     {
         [ForeignKey("Character")]
         public int Id { get; set; }
-        public double Tall { get; set; }
+        public int Tall { get; set; }
         public string BodyType { get; set; }
         public string Description { get; set; }
         public  int BreedId { get; set; }
